+++ packages/moengage_flutter/moengage_flutter_android/android/src/main/kotlin/com/moengage/flutter/Constants.kt
@@ -53,3 +53,8 @@ const val METHOD_NAME_SHOW_NUDGE = "showNudge"
 // Self Handled InApps
 const val METHOD_NAME_SELF_HANDLED_IN_APPS = "selfHandledInApps"
 const val ERROR_CODE_SELF_HANDLED_IN_APPS = "SELF_HANDLED_IN_APPS_ERROR"
+
+// Identify User
+const val METHOD_NAME_IDENTIFY_USER = "identifyUser"
+const val METHOD_NAME_GET_USER_IDENTITIES = "getUserIdentities"
+const val ERROR_CODE_GET_USER_IDENTITIES = "GET_USER_IDENTITIES_ERROR"