+++ packages/moengage_flutter/moengage_flutter_platform_interface/test/data_provider/data_provider.dart
@@ -219,3 +219,8 @@ final SelfHandledCampaignsData selfHandledCampaigns =
           Rules(screenName: 'Dashboard', context: ['section_1', 'section_2'])),
       Platforms.android),
 ]);
+
+const dummyAppId = 'Dummy_App_ID';
+const accountMetaPayload = {
+  keyAccountMeta: {keyAppId: dummyAppId}
+};